+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Version 0.3.0 / Unreleased
 
  * [FEATURE] De-/compress using libbz2 via FFI
+ * [FEATURE] De-/compress using Commons Compress on JRuby
  * [ENHANCEMENT] Performance improvements for the pure Ruby implementation
 
 ## Version 0.2.0 / 2011-11-11